backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Limiter==3.6.0

pymongo==4.6.1
python-dotenv==1.0.1

Werkzeug==3.0.1
